requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-appconf==1.0.5
django-imagekit==4.1.0
django-phonenumber-field==6.1.0
fontawesome-free==5.15.4
phonenumberslite==8.12.44
pilkit==2.0
Pillow==9.0.1
six==1.16.0
sqlparse==0.4.2
psycopg2-binary